fix: to add uniquevalue legend data for stac cog
